refactor: Update migrations to add nullable fields and modify table structure
This commit updates the database migrations to add nullable fields and modify the table structure. Specifically, it adds the 'uuid' column to the 'media' table, changes the 'user_id' column in the 'peserta' table to 'unsignedBigInteger', and adds the 'referensi' column to the 'registrasi' table. These changes improve the flexibility and functionality of the database schema, allowing for better data management and organization.

+++ database/migrations/2024_05_07_011103_create_peserta_mbkm_table.php
@@ -6,12 +6,12 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    
+
     public function up(): void
     {
         Schema::create('peserta', function (Blueprint $table) {
             $table->id();
-            $table->integer('user_id')->unsigned();
+            $table->unsignedBigInteger('user_id')->unsigned();
             $table->string('nama');
             $table->string('nim')->unique();
             $table->string('email')->unique();
